+++ client/src/features/question-composer/CoverLetterDetail.tsx
