+++ client/src/features/question-composer/index.ts
@@ -1,0 +1,1 @@
+export { default as QuestionComposer } from './QuestionComposer';
